+++ kamehouse-shell/bin/kamehouse/kamehouse-desktop.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+main() {
+  # Execute the latest deployed version of kamehouse-desktop
+  ${HOME}/programs/kamehouse-desktop/bin/kamehouse-desktop.sh "$@"
+}
+
+main "$@"