+++ kamehouse-shell/bin/kamehouse/kamehouse-snape.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+main() {
+  # Execute the latest deployed version of kamehouse-snape
+  ${HOME}/programs/kamehouse-snape/bin/kamehouse-snape.sh "$@"
+}
+
+main "$@"