+++ kamehouse-snape/bin/kamehouse-desktop/kamehouse-desktop.py
@@ -1,0 +1,4 @@
+import sys
+
+print("kamehouse-desktop started! with args:")
+print(sys.argv[1:])